Clean up object hierarchy
