+++ itsim/machine/process_management/daemon.py
@@ -2,7 +2,8 @@ from .__init__ import _Service, _Thread
 
 from typing import Callable
 
-from itsim import _Node, ITObject
+from itsim import ITObject
+from itsim.machine import _Node
 from itsim.simulator import Simulator
